Move unknown function logic into dedicated file
unknown_function_handling.c now holds all of the unknown function handling
logic.
